+++ cmp/pinpad/PinPadModel.js
@@ -1,7 +1,11 @@
-import FastClick from '@onsenui/fastclick';
+/*
+ * This file belongs to Hoist, an application development toolkit
+ * developed by Extremely Heavy Industries (www.xh.io | info@xh.io)
+ *
+ * Copyright © 2020 Extremely Heavy Industries Inc.
+ */
 import {HoistModel} from '@xh/hoist/core';
 import {action, bindable, computed, observable} from '@xh/hoist/mobx';
-import {createObservableRef} from '@xh/hoist/utils/react';
 import {times} from 'lodash';
 
 @HoistModel
@@ -16,12 +20,9 @@ export class PinPadModel {
     /** @member {string} */
     @bindable errorText;
 
-    ref = createObservableRef();
-
     @observable _enteredDigits;
     _deleteWasLast = false;
     _pinLength;
-
 
     /**
      * @param {Object} [c] - configuration object.
@@ -39,13 +40,6 @@ export class PinPadModel {
 
         this._pinLength = pinLength;
         this._enteredDigits = [];
-
-        this.addReaction({
-            track: () => this.ref.current,
-            run: (current) => {
-                if (current) FastClick.attach(current);
-            }
-        });
     }
 
     //-------------------